Move errors to signac.errors
* Deprecation warning to error modules
* core, contrib, common error modules have deprecation warning
