+++ signac/contrib/import_export.py
@@ -16,8 +16,8 @@ from string import Formatter
 from ..common import six
 from ..common.tempdir import TemporaryDirectory
 from ..core import json
-from .errors import StatepointParsingError
-from .errors import DestinationExistsError
+from ..errors import StatepointParsingError
+from ..errors import DestinationExistsError
 from .utility import _mkdir_p
 
 import logging
